add interface to separate customer and offer database methods
