+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/offers/OfferDbConnector.groovy
@@ -30,14 +30,16 @@ class OfferDbConnector implements CreateOfferDataSource{
 
     DatabaseSession session
     ConnectionProvider connectionProvider
+    OfferToCustomerGateway offerToCustomerGateway
 
     private static final Logger LOG = LogManager.getLogger(OfferDbConnector.class)
 
     private static final String OFFER_INSERT_QUERY = "INSERT INTO offer (modificationDate, expirationDate, customerId, projectManagerId, projectTitle, projectDescription, totalPrice, customerAffiliationId)"
 
 
-    OfferDbConnector(DatabaseSession session){
+    OfferDbConnector(DatabaseSession session, OfferToCustomerGateway offerToCustomerGateway){
         this.session = session
+        this.offerToCustomerGateway = offerToCustomerGateway
     }
 
     /**
@@ -57,10 +59,12 @@ class OfferDbConnector implements CreateOfferDataSource{
 
             connection.withCloseable {it ->
                 try {
-                    int projectManagerId = getPersonId(offer.projectManager)
-                    int customerId = getPersonId(offer.customer)
-                    int affiliationId = getAffiliationId(offer.selectedCustomerAffiliation)
+                    int projectManagerId = offerToCustomerGateway.getPersonId(connection, offer.projectManager)
+                    int customerId = offerToCustomerGateway.getPersonId(connection, offer.customer)
+                    int affiliationId = offerToCustomerGateway.getAffiliationId(connection,offer.selectedCustomerAffiliation)
+
                     createOffer(it, offer, projectManagerId, customerId, affiliationId)
+
                     connection.commit()
                 } catch (Exception e) {
                     log.error(e.message)
@@ -80,7 +84,7 @@ class OfferDbConnector implements CreateOfferDataSource{
     }
 
     /**
-     *
+     * The
      * @param offer
      * @return
      */
@@ -111,61 +115,5 @@ class OfferDbConnector implements CreateOfferDataSource{
         return generatedKeys[0]
     }
 
-    /**
-     * Searches for a person and returns the matching ID from the person table
-     * @param person The database entry is searched for a person
-     * @return the ID of the database entry matching the person
-     */
-    int getPersonId(Person person) {
-        String query = "SELECT id FROM person WHERE first_name = ? AND last_name = ? AND email = ?"
-        Connection connection = connectionProvider.connect()
 
-        int personId = -1
-
-        connection.withCloseable {
-            def statement = connection.prepareStatement(query)
-            statement.setString(1, person.firstName)
-            statement.setString(2, person.lastName)
-            statement.setString(3, person.emailAddress)
-
-            ResultSet result = statement.executeQuery()
-            while (result.next()){
-                personId = result.getInt(1)
-            }
-        }
-        return personId
-    }
-
-    //todo the same method as in customerDbconnector
-    int getAffiliationId(Affiliation affiliation) {
-        String query = "SELECT id FROM affiliation WHERE organization=? " +
-                "AND address_addition=? " +
-                "AND street=? " +
-                "AND postal_code=? " +
-                "AND city=?"
-
-        Connection connection = connectionProvider.connect()
-
-        List<Integer> affiliationIds = []
-
-        def statement = connection.prepareStatement(query)
-        statement.setString(1, affiliation.organisation)
-        statement.setString(2, affiliation.addressAddition)
-        statement.setString(3, affiliation.street)
-        statement.setString(4, affiliation.postalCode)
-        statement.setString(5, affiliation.city)
-
-        ResultSet rs = statement.executeQuery()
-        while (rs.next()) {
-            affiliationIds.add(rs.getInt(1))
-        }
-
-        if(affiliationIds.size() > 1) {
-            throw new DatabaseQueryException("More than one entry found for $affiliation.")
-        }
-        if (affiliationIds.empty) {
-            throw new DatabaseQueryException("No matching affiliation found for $affiliation.")
-        }
-        return affiliationIds[0]
-    }
 }
